requirement.txt
django
python-dotenv
google-genai
google
